==> Expert_Follow_Graph/3_scrape_followers_following.ipynb <==
import pywren, json, requests, time, bs4
import pandas as pd
from re import findall
import numpy as np

data = pd.read_csv('complete_data.csv')
mobile = "https://mobile.twitter.com"
expert_data = json.load(open('expert_network.json'))
tar_accounts = list(set(data['twitter'])-expert_data.keys()-{'Twitter', '', '-1'})
print(len(tar_accounts)), tar_accounts

## for each page, extract followers/ following
def Follow(response):
    soup = bs4.BeautifulSoup(response, "html.parser")
    follow = soup.find_all("td", "info fifty screenname")
    cursor = soup.find_all("div", "w-button-more")
    try: cursor = findall(r'cursor=(.*?)">', str(cursor))[0]
    except IndexError: pass  
    follow = [i.find('a')['name'] for i in follow]
    return follow, cursor

## given a user, loop through its followers/ following pages using cursors
def scrape_fols(username, fol):
    mobile = "https://mobile.twitter.com"
    url = f"{mobile}/{username}/{fol}?lang=en"
    init = '-1'
    follower_lst = []
    count = 0
    
    while(len(init)!=0):
        if init == '-1': tar_url = url
        else: tar_url = url + f"&cursor={init}"
        r = requests.get(tar_url)
        
        for i in range(5):
            if r: break
            else: 
                time.sleep(3)
                r = requests.get(tar_url).text
                
        if not r: break
        followers, init = Follow(r.text)
        follower_lst.extend(followers)
        print(len(follower_lst), end = '\r')
    return follower_lst

def scrape_user(username):
    expert = dict()
    expert["followers"] = scrape_fols(username, fol = 'followers')
    expert["following"] = scrape_fols(username, fol = 'following')
    return {username: expert}

pwex = pywren.default_executor(job_max_runtime = 500)
futures = pwex.map(scrape_user, tar_accounts)
print(len(futures))

not_dones = [1]
while len(not_dones) != 0:
    dones, not_dones = pywren.wait(futures, pywren.ANY_COMPLETED)
    print(len(dones), len(not_dones), end = '\r')

#expert_data = dict()
for num, item in enumerate(dones):
    try: expert_data.update(item.result())
    except (IndexError, AttributeError): continue
len(expert_data.keys())

json.dump(expert_data, open('expert_network_v2.json', 'w'))

extra = set(tar_accounts) - expert_data.keys()
print(len(extra))
for user in extra:
    if user == 'MarshallJenA': continue
    print(user)
    add = scrape_user(user)
    expert_data.update(add)
    print(len(expert_data[user]['followers']), 
          len(expert_data[user]['following']))
    json.dump(expert_data, open('expert_network.json', 'w'))

extra = set(tar_accounts) - expert_data.keys()
extra
