==> Expert_Follow_Graph/9_basic_demograph.ipynb <==
import sklearn, json
import pandas as pd
import seaborn as sns
import numpy as np
import matplotlib.pyplot as plt
from collections import Counter

info = pd.read_csv('expert_twitters_complete.csv', encoding = "ISO-8859-1")
info2 = pd.read_csv('expert_bio_complete.csv')
info = info.drop(['href'], axis=1)
info2 = info2.drop(['name', 'title', 'institution','education', 'experience'], axis=1)

data = info.merge(info2, on = 'expert_id', how='left')
data = data.drop_duplicates()
data.to_csv('complete_data.csv', index=False)

## Expert Percentage

data = pd.read_csv('complete_data.csv')
data = data[~data['type'].isin(['N', 'other'])]

len(set(data['expert_id'])), len(set(data[data['twitter']!='-1']['expert_id']))

for ins in ['aei', 'brookings', 'cap', 'heritage']:
    cut = data[data['institution']==ins]
    a = len(set(cut['expert_id']))
    b = len(set(cut[cut['twitter']!='-1']['expert_id']))
    print(ins, b/a)

Counter(data['is_phd'])

413/(413+508)

np.corrcoef(info['num_following'], info['num_followers'])
